feat: Added tests for getAllMarkdownFiles

+++ __tests__/getAllMarkdownFiles.test.ts
@@ -22,7 +22,8 @@ describe('getAllMarkdownFiles', () => {
 
         fs.writeFileSync(path.join(testDir, 'file1.md'), '# File 1');
         fs.writeFileSync(path.join(testDir, 'file2.txt'), 'This is not markdown.');
-        fs.writeFileSync(path.join(testDir, 'README.md'), '# README (uppercase extension)');
+        fs.writeFileSync(path.join(testDir, 'README.md'), '# README (uppercase)');
+        fs.writeFileSync(path.join(testDir, 'file4'), '# File 4 (without .md extension)');
 
         const subDir = path.join(testDir, 'subdir');
         fs.mkdirSync(subDir);
@@ -32,10 +33,12 @@ describe('getAllMarkdownFiles', () => {
 
     // Removes temporary directory after each test
     afterEach(() => {
-        fs.rmSync(testDir, { recursive: true, force: true });
+        if (fs.existsSync(testDir)) {
+            fs.rmSync(testDir, { recursive: true, force: true });
+        }
     });
 
-    test('Return all markdown files by default', () => {
+    test('Returns all markdown files by default', () => {
         const result = getAllMarkdownFiles(testDir);
 
         const expectedFiles = [
@@ -49,7 +52,7 @@ describe('getAllMarkdownFiles', () => {
         expect(result.sort()).toEqual(expectedFiles.sort());
     });
 
-    test('Ignore specified markdown files (case-insensitive)', () => {
+    test('Ignores specified markdown files (case-insensitive)', () => {
         const result = getAllMarkdownFiles(
             testDir,
             { ignoreFilenames: ['index.md', 'readme.md']}
@@ -64,12 +67,31 @@ describe('getAllMarkdownFiles', () => {
         expect(result.sort()).toEqual(expectedFiles.sort());
     });
 
-    test('Work with an empty directory', () => {
+    test('Does not return files without .md extension', () => {
+        const result = getAllMarkdownFiles(testDir);
+
+        expect(result).not.toContain(path.join(testDir, 'file4'));
+
+    });
+
+    test('Works with an empty directory', () => {
         const emptyDir = path.join(testDir, 'empty');
         fs.mkdirSync(emptyDir);
 
         const result = getAllMarkdownFiles(emptyDir);
         expect(result).toEqual([]);
+    });
+
+    test('Handles non-existent directory', () => {
+        const nonExistentDir = path.join(testDir, 'does-not-exist');
+
+        if (fs.existsSync(nonExistentDir)) {
+            fs.rmSync(nonExistentDir, { recursive: true, force: true });
+        }
+
+        expect(() => getAllMarkdownFiles(nonExistentDir)).toThrow(
+            new Error(`Path "${nonExistentDir}" is not a valid directory.`)
+        );
     });
 
 });
